add battle room

+++ backend/go/api/battle/input.go
@@ -3,7 +3,7 @@ package battle
 type Input struct {
 	Time int64
 	Key  Keys
-	Side int
+	Side int8
 }
 
 type Keys struct {
